+++ packages/messenger-widget/src/hooks/topLevelAlias/nameService/Genome.ts
@@ -1,5 +1,5 @@
 import { ethers } from 'ethers';
-import { ITLDResolver, TLDResolver } from './TLDResolver';
+import { ITLDResolver } from './ITLDResolver';
 import { createWeb3Name } from '@web3-name-sdk/core';
 
 const TOP_LEVEL_DOMAIN = '.gno';
