+++ packages/messenger-widget/src/hooks/topLevelAlias/nameService/ITLDResolver.ts
@@ -1,0 +1,13 @@
+export interface ITLDResolver {
+    isResolverForTldName(ensName: string): Promise<boolean>;
+    isResolverForAliasName(
+        ensName: string,
+        foreinAliasName?: string,
+    ): Promise<boolean>;
+    resolveAliasToTLD(
+        ensName: string,
+        //In case a foreign name needs to be resolved, the tldName is provided by the stroage, using this optional parameter
+        foreignTldName?: string,
+    ): Promise<string>;
+    resolveTLDtoAlias(ensName: string): Promise<string>;
+}
